feat: delete old electric proxy
