+++ apps/backend/package.json
@@ -43,7 +43,7 @@
 		"@opentelemetry/sdk-metrics": "^2.2.0",
 		"@opentelemetry/sdk-trace-node": "^2.2.0",
 		"@opentelemetry/sdk-trace-web": "^2.2.0",
-		"@workos-inc/node": "^7.73.0",
+		"@workos-inc/node": "^7.74.2",
 		"drizzle-orm": "^0.44.7",
 		"effect": "catalog:effect",
 		"jose": "^6.1.2",
