--- apps/electric-proxy-bun/package.json
@@ -1,28 +1,0 @@
-{
-	"name": "@hazel/electric-proxy-bun",
-	"version": "0.0.0",
-	"private": true,
-	"type": "module",
-	"scripts": {
-		"dev": "bun run --hot src/index.ts",
-		"start": "bun run src/index.ts",
-		"typecheck": "tsc --noEmit"
-	},
-	"devDependencies": {
-		"@types/bun": "1.3.3",
-		"typescript": "^5.9.3"
-	},
-	"dependencies": {
-		"@effect/opentelemetry": "catalog:effect",
-		"@effect/platform": "catalog:effect",
-		"@effect/platform-bun": "catalog:effect",
-		"@electric-sql/client": "^1.1.5",
-		"@hazel/db": "workspace:*",
-		"@hazel/schema": "workspace:*",
-		"@opentelemetry/exporter-trace-otlp-http": "^0.57.0",
-		"@opentelemetry/sdk-trace-node": "^2.2.0",
-		"@workos-inc/node": "^7.73.0",
-		"effect": "catalog:effect",
-		"jose": "^6.1.2"
-	}
-}
