+++ apps/electric-proxy/package.json
@@ -2,27 +2,27 @@
 	"name": "@hazel/electric-proxy",
 	"version": "0.0.0",
 	"private": true,
+	"type": "module",
 	"scripts": {
-		"deploy": "wrangler deploy",
-		"devxd": "wrangler dev --port 8184",
-		"start": "wrangler dev",
-		"test": "vitest",
-		"cf-typegen": "wrangler types"
+		"dev": "bun run --hot src/index.ts",
+		"start": "bun run src/index.ts",
+		"typecheck": "tsc --noEmit"
 	},
 	"devDependencies": {
-		"typescript": "^5.9.3",
-		"vitest": "^4.0.13",
-		"wrangler": "^4.50.0"
+		"@types/bun": "1.3.3",
+		"typescript": "^5.9.3"
 	},
 	"dependencies": {
+		"@effect/opentelemetry": "catalog:effect",
 		"@effect/platform": "catalog:effect",
-		"@electric-sql/client": "^1.1.5",
+		"@effect/platform-bun": "catalog:effect",
+		"@electric-sql/client": "^1.2.0",
 		"@hazel/db": "workspace:*",
 		"@hazel/schema": "workspace:*",
-		"@workos-inc/node": "^7.73.0",
-		"drizzle-orm": "^0.44.7",
+		"@opentelemetry/exporter-trace-otlp-http": "^0.208.0",
+		"@opentelemetry/sdk-trace-node": "^2.2.0",
+		"@workos-inc/node": "^7.74.2",
 		"effect": "catalog:effect",
-		"jose": "^6.1.2",
-		"postgres": "^3.4.7"
+		"jose": "^6.1.2"
 	}
 }
