+++ apps/electric-proxy/src/auth/bot-auth.ts
