+++ apps/electric-proxy/src/config.ts
