+++ apps/electric-proxy/src/proxy/electric-client.ts
