+++ apps/electric-proxy/src/tables/bot-tables.ts
