--- apps/electric-proxy/wrangler.jsonc
@@ -1,38 +1,0 @@
-/**
- * For more details on how to configure Wrangler, refer to:
- * https://developers.cloudflare.com/workers/wrangler/configuration/
- */
-{
-	"$schema": "node_modules/wrangler/config-schema.json",
-	"name": "electric-proxy",
-	"main": "src/index.ts",
-	"compatibility_date": "2025-11-19",
-	"compatibility_flags": [
-		"nodejs_compat"
-	],
-	"observability": {
-		"enabled": true
-	},
-	"routes": [
-		{
-			"pattern": "electric.hazel.sh",
-			"custom_domain": true
-		}
-	],
-	"hyperdrive": [
-		{
-			"binding": "HYPERDRIVE",
-			"id": "8f2b1e43f91746c494480d2516720909",
-			"localConnectionString": "postgresql://user:password@localhost:5432/app"
-		}
-	],
-	"placement": {
-		"mode": "smart"
-	},
-	"vars": {
-		"ELECTRIC_URL": "https://api.electric-sql.cloud",
-		"ELECTRIC_SOURCE_ID": "80422006-0448-4af1-9815-5d29c2dd5fba",
-		"WORKOS_CLIENT_ID": "client_01HGZR2D24QXAZP47AWRENTA8M",
-		"ALLOWED_ORIGIN": "https://app.hazel.sh"
-	}
-}
