+++ packages/backend-core/package.json
@@ -16,7 +16,7 @@
 		"@effect/platform": "catalog:effect",
 		"@hazel/db": "workspace:*",
 		"@hazel/domain": "workspace:*",
-		"@workos-inc/node": "^7.73.0",
+		"@workos-inc/node": "^7.74.2",
 		"drizzle-orm": "^0.44.7",
 		"effect": "catalog:effect"
 	},
