+++ packages/effect-electric-db-collection/package.json
@@ -17,7 +17,7 @@
 	},
 	"dependencies": {
 		"@effect-atom/atom": "catalog:effect",
-		"@electric-sql/client": "^1.1.5",
+		"@electric-sql/client": "^1.2.0",
 		"@standard-schema/spec": "^1.0.0",
 		"@tanstack/db": "0.5.0",
 		"@tanstack/electric-db-collection": "0.2.0",
